// frontend/messages.d.ts
export type msg = "light_giant_shell_attend"|"stale_minor_cuckoo_type"|"minor_orange_hawk_honor"|"nice_mushy_ox_vent"|"blue_alert_scallop_hope"|"careful_awful_myna_scold"|"less_sleek_nils_read"|"fancy_known_peacock_pray";
export const langs = ["en","es","de","fa","fr","gu","hi","ja","ko","pl","pt-br","ro","ru","tr","zh-cn","zh-hk","zh-tw"] as const;